Renaming collections to cobbler_collections
Due to a naming conflict with a python core module I had to rename the module to cobbler_collections.
