+++ cobbler/cobbler_collections/repos.py
@@ -1,5 +1,5 @@
 """
-Copyright 2008-2009, Red Hat, Inc and Others
+Copyright 2006-2009, Red Hat, Inc and Others
 Michael DeHaan <michael.dehaan AT gmail>
 
 This program is free software; you can redistribute it and/or modify
@@ -18,56 +18,64 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 02110-1301  USA
 """
 
-from cobbler.actions import litesync
-from cobbler.collections import collection
-from cobbler.items import system as system
+import os.path
+
+from cobbler.cobbler_collections import collection
+from cobbler.items import repo as repo
 from cobbler import utils
 from cobbler.cexceptions import CX
 from cobbler.utils import _
 
 
-class Systems(collection.Collection):
+class Repos(collection.Collection):
     """
-    Systems are hostnames/MACs/IP names and the associated profile
-    they belong to.
+    Repositories in cobbler are way to create a local mirror of a yum repository.
+    When used in conjunction with a mirrored distro tree (see "cobbler import"),
+    outside bandwidth needs can be reduced and/or eliminated.
     """
 
     def collection_type(self):
-        return "system"
+        return "repo"
 
-    def factory_produce(self, collection_mgr, item_dict):
+    def factory_produce(self, config, item_dict):
         """
         Return a Distro forged from item_dict
         """
-        new_system = system.System(collection_mgr)
-        new_system.from_dict(item_dict)
-        return new_system
+        new_repo = repo.Repo(config)
+        new_repo.from_dict(item_dict)
+        return new_repo
 
     def remove(self, name, with_delete=True, with_sync=True, with_triggers=True, recursive=False, logger=None):
         """
         Remove element named 'name' from the collection
         """
+        # NOTE: with_delete isn't currently meaningful for repos
+        # but is left in for consistancy in the API.  Unused.
         name = name.lower()
         obj = self.find(name=name)
-
         if obj is not None:
-
             if with_delete:
                 if with_triggers:
-                    utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/delete/system/pre/*", [], logger)
-                if with_sync:
-                    lite_sync = litesync.CobblerLiteSync(self.collection_mgr, logger=logger)
-                    lite_sync.remove_single_system(name)
+                    utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/delete/repo/pre/*", [], logger)
+
             self.lock.acquire()
             try:
                 del self.listing[name]
             finally:
                 self.lock.release()
             self.collection_mgr.serialize_delete(self, obj)
+
             if with_delete:
                 if with_triggers:
-                    utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/delete/system/post/*", [], logger)
+                    utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/delete/repo/post/*", [], logger)
                     utils.run_triggers(self.collection_mgr.api, obj, "/var/lib/cobbler/triggers/change/*", [], logger)
+
+                # FIXME: better use config.settings() webdir?
+                path = "/var/www/cobbler/repo_mirror/%s" % obj.name
+                if os.path.exists("/srv/www/"):
+                    path = "/srv/www/cobbler/repo_mirror/%s" % obj.name
+                if os.path.exists(path):
+                    utils.rmtree(path)
 
             return
