+++ cobbler/web/templatetags/site.py
@@ -1,6 +1,6 @@
 from builtins import object
 from django import template
-from collections import OrderedDict
+from cobbler_collections import OrderedDict
 
 register = template.Library()
